Mis cambios locales

+++ controllers/clienteController.js
@@ -1,7 +1,9 @@
 const Cliente = require('../models/cliente');
+const {validationResult} = require('express-validator');
 
 
 const getClientes = async (req, res) => {
+  
   try {
     const clientes = await Cliente.find();
     res.json(clientes);
@@ -11,6 +13,10 @@ const getClientes = async (req, res) => {
 };
 
 const getClienteById = async (req, res) => {
+  const errors = validationResult(req);
+  if(!errors.isEmpty()) {
+      return res.status(400).json({errors: errors.array()})
+  }
   try {
     const { id } = req.params;
     const cliente = await Cliente.findById(id);
@@ -21,6 +27,10 @@ const getClienteById = async (req, res) => {
 };
 
 const postCliente = async (req, res) => {
+  const errors = validationResult(req);
+      if(!errors.isEmpty()) {
+          return res.status(400).json({errors: errors.array()})
+      }
   try {
     const { body } = req;
     const cliente = new Cliente(body);
@@ -32,12 +42,16 @@ const postCliente = async (req, res) => {
 };
 
 const putCliente = async (req, res) => {
+  const errors = validationResult(req);
+  if(!errors.isEmpty()) {
+      return res.status(400).json({errors: errors.array()})
+  }
   try {
-    const { documento, nombres, apellidos, correo, telefono, estado } = req.body;
+    const { documento, nombre, apellido, correo, telefono, estado } = req.body;
     const id = req.params.id;
     await Cliente.findOneAndUpdate(
       { _id: id },
-      { documento, nombres, apellidos, correo, telefono, estado }
+      { documento, nombre, apellido, correo, telefono, estado }
     );
     res.json({ msg: 'Cliente actualizado exitosamente' });
   } catch (error) {
@@ -46,6 +60,10 @@ const putCliente = async (req, res) => {
 };
 
 const deleteCliente = async (req, res) => {
+  const errors = validationResult(req);
+  if(!errors.isEmpty()) {
+      return res.status(400).json({errors: errors.array()})
+  }
   try {
     const id = req.params.id;
     await Cliente.findOneAndDelete({ _id: id });
